+++ clang/lib/Sema/SemaDecl.cpp
@@ -7627,6 +7627,16 @@ static bool isMainVar(DeclarationName Name, VarDecl *VD) {
           VD->isExternC());
 }
 
+void Sema::DiagnoseWeakPointerAuthenticationSchema(VarDecl *VD) {
+  if (Context.isPointerAuthenticationAvailable() &&
+      VD->isFunctionPointerType() && !VD->isExternallyVisible()) {
+    PointerAuthQualifier Q = VD->getType().getQualifiers().getPointerAuth();
+    if (!Q || (!Q.isAddressDiscriminated() && Q.getExtraDiscriminator() == 0)) {
+      Diag(VD->getLocation(), diag::warn_ptrauth_weak_schema) << VD << !Q;
+    }
+  }
+}
+
 NamedDecl *Sema::ActOnVariableDeclarator(
     Scope *S, Declarator &D, DeclContext *DC, TypeSourceInfo *TInfo,
     LookupResult &Previous, MultiTemplateParamsArg TemplateParamLists,
@@ -8356,6 +8366,8 @@ NamedDecl *Sema::ActOnVariableDeclarator(
     checkDLLAttributeRedeclaration(*this, Prev, NewVD, IsMemberSpecialization,
                                    D.isFunctionDefinition());
   }
+
+  DiagnoseWeakPointerAuthenticationSchema(NewVD);
 
   if (NewTemplate) {
     if (NewVD->isInvalidDecl())